SCRUM-33: refactor file structure
