--- src/main/java/com/gmail/ypon2003/marketplacebackend/dto/PersonAsSellerDTO.java
@@ -1,13 +1,0 @@
-package com.gmail.ypon2003.marketplacebackend.dto;
-
-import lombok.Builder;
-
-@Builder
-public record PersonAsSellerDTO(
-        String name,
-        String lastName,
-        String password,
-        String role,
-        String email,
-        String phoneNumber) {
-}
