--- src/main/java/com/gmail/ypon2003/marketplacebackend/dto/ProductDTO.java
@@ -1,18 +1,0 @@
-package com.gmail.ypon2003.marketplacebackend.dto;
-
-import lombok.Builder;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-@Builder
-public record ProductDTO(
-        String name,
-        String description,
-        BigDecimal price,
-        String measurement,
-        LocalDateTime createAt,
-        String infoSeller,
-        Long personId
-) {
-}
