--- src/main/java/com/gmail/ypon2003/marketplacebackend/services/ProductService.java
@@ -1,105 +1,0 @@
-package com.gmail.ypon2003.marketplacebackend.services;
-
-import com.gmail.ypon2003.marketplacebackend.dto.ProductDTO;
-import com.gmail.ypon2003.marketplacebackend.models.Person;
-import com.gmail.ypon2003.marketplacebackend.models.Product;
-import com.gmail.ypon2003.marketplacebackend.repositories.PersonRepository;
-import com.gmail.ypon2003.marketplacebackend.repositories.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-
-/**
- * @author uriiponomarenko 28.05.2024
- */
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-@Slf4j
-public class ProductService {
-
-    private final ProductRepository productRepository;
-    private final PersonRepository personRepository;
-
-    public List<Product> findAll() {
-        return productRepository.findAll();
-    }
-
-    @Transactional
-    public Product save(ProductDTO productDTO) {
-        try {
-
-            Person person = personRepository.findById(productDTO.personId())
-                    .orElseThrow(() -> new RuntimeException("Person with id " +
-                    productDTO.personId() + " not found"));
-
-            Product product = Product.builder()
-                    .createAt(LocalDateTime.now())
-                    .name(productDTO.name())
-                    .measurement(productDTO.measurement())
-                    .description(productDTO.description())
-                    .price(productDTO.price())
-                    .infoSeller(productDTO.infoSeller())
-                    .person(person)
-                    .build();
-            return productRepository.save(product);
-        } catch (Exception e) {
-            log.error("Failed to save product", e);
-            throw new RuntimeException("Failed to save product", e);
-        }
-    }
-
-    public Optional<Product> showProduct(long id) {
-        return productRepository.findById(id);
-    }
-
-    @Transactional
-    public void updateProduct(long id, ProductDTO productDTO) {
-        Optional<Product> updateToBeProduct = showProduct(id);
-        if (updateToBeProduct.isPresent()) {
-            Product product = updateToBeProduct.get();
-            product.setName(productDTO.name());
-
-            if(productDTO.createAt() != null) {
-                product.setCreateAt(productDTO.createAt());
-            }
-
-            product.setMeasurement(productDTO.measurement());
-            product.setDescription(productDTO.description());
-            product.setPrice(productDTO.price());
-            product.setInfoSeller(productDTO.infoSeller());
-
-            productRepository.save(product);
-        }
-    }
-
-    @Transactional
-    public void deleteProduct(long id) {
-        productRepository.deleteById(id);
-    }
-
-    public List<Product> searchProductsByName(String name) {
-        return productRepository.findByNameContaining(name);
-    }
-
-    public List<Product> getProductsSortedByPrice() {
-        return productRepository.findAll(Sort.by(Sort.Direction.ASC, "price"));
-    }
-
-    public List<Product> getProductsSortedByDate() {
-        return productRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt"));
-    }
-
-    public Page<Product> getProductsPage(Pageable pageable) {
-        return productRepository.findAll(pageable);
-    }
-
-}
