+++ src/main/java/com/khutircraftubackend/MarketPlaceBackendApplication.java
@@ -1,9 +1,11 @@
-package com.gmail.ypon2003.marketplacebackend;
+package com.khutircraftubackend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
+@EnableAsync
 public class MarketPlaceBackendApplication {
 
     public static void main(String[] args) {
